Separate user searching logic as Query model
