+++ src/main/kotlin/com/keyskey/ktknowledge/handlers/graphql/queries/UserQuery.kt
@@ -5,22 +5,22 @@ import com.expediagroup.graphql.server.operations.Query
 import com.keyskey.ktknowledge.handlers.graphql.types.UserType
 import com.keyskey.ktknowledge.handlers.graphql.types.toUserType
 import com.keyskey.ktknowledge.handlers.graphql.utils.toInt
-import com.keyskey.ktknowledge.repositories.UserRepository
+import com.keyskey.ktknowledge.usecases.user.UserSearcher
 import org.springframework.stereotype.Component
 
 @Component
-class UserQuery(private val userRepository: UserRepository): Query {
-    fun users(): List<UserType> {
-        return userRepository
-            .findAll()
+class UserQuery(private val userSearcher: UserSearcher): Query {
+    fun newUsers(): List<UserType> {
+        return userSearcher
+            .findNewUsers()
             .map { it.toUserType() }
     }
 
     fun user(id: ID): UserType? {
         val intId = id.toInt()
 
-        return userRepository
-            .findById(intId)
+        return userSearcher
+            .findOne(intId)
             ?.toUserType()
     }
 }
